backend/pyproject.toml
[project]
name = "qr-ems-backend"
version = "0.1.0"
description = "QR-EMS Backend API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.0",
    "djangorestframework>=3.15",
    "drf-spectacular>=0.27",
    "psycopg[binary]>=3.1",
    "python-decouple>=3.8",
    "qrcode>=7.4",
    "rdflib>=7.0",
    "pillow>=10.0.0",
    "gunicorn>=21.2.0",
    "dj-database-url>=3.0.1",
    "djangorestframework-simplejwt>=5.5.1",
    "google-auth>=2.43.0",
    "requests>=2.32.5",
    "django-cors-headers>=4.9.0",
    "django-storages[s3]>=1.14.6",
    "boto3>=1.42.25",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "coverage>=7.0",
    "ruff>=0.9.0",
]

[tool.coverage.run]
source = ["."]
omit = [
    "*/migrations/*",
    "*/tests.py",
    "*/test_*.py",
    "manage.py",
    "config/*",
    "fix_borrowed_data.py",
    "generate_test_data.py",
]

[tool.coverage.report]
show_missing = true
skip_empty = true

[tool.ruff]
# Target version
target-version = "py312"
line-length = 88
# Exclude a variety of commonly ignored directories.
exclude = [
    "migrations",
    "__pycache__",
    ".venv",
    ".git",
    "tests.py",
]

[tool.ruff.lint]
# Enable Pyflakes (F) and Pycodestyle (E, W) codes by default.
select = ["E", "F", "I", "N", "UP", "B", "A", "C4", "SIM", "ARG"]
ignore = ["E501"] # Ignore line too long

[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
